changed host to be Docker rather than directly on top of Vagrant;
This is so that when the architecture becomes more complex (e.g. introduction of database), we're not having to either run multiple VMs or more heavily customize the one we have. It also enables us to use vanilla, public images rather than install software on the VM. Finally, this should give more flexibility who want to use the Docker tool-chain and not use Vagrant at all